requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.4.0+cu121
torchvision==0.19.0+cu121

diffusers>=0.30.0
transformers>=4.44.1
sentencepiece>=0.2.0
accelerate>=0.33.0
protobuf>=5.0
optimum-quanto==0.2.4